Python fájlok frissítve

+++ python/lux.py
@@ -24,7 +24,7 @@ worksheet.write('C1','Látható fény(lux)',bold)
 worksheet.write('D1','Infravörös(lux)',bold)
 #a szenzor adatainak beszerzése
 frek=int(sys.argv[2])
-nagyt=int(1/frek)
+nagyt=1/frek
 idotartam=int(sys.argv[3])
 #ciklushoz kapcsolódó változók deklarálása
 ciklus=0
@@ -32,7 +32,7 @@ ciklusseged=0
 row=1
 column=0
 while ciklus<idotartam:
-    f.write("%d#" %ciklusseged*nagyt)
+    f.write("%d#" %ciklusseged*int(nagyt))
     f.write("%d#" %sensor.lux)
     f.write("%d#" %sensor.visible)
     f.write("%d##\r\n"%sensor.infrared)
