+++ python/press.py
@@ -21,7 +21,7 @@ worksheet.write('B1','Magasság(m)',bold)
 worksheet.write('C1','Nyomás(Pa)',bold)
 #a szenzor adatainak beszerzése
 frek=int(sys.argv[2])
-nagyt=int(1/frek)
+nagyt=1/frek
 idotartam=int(sys.argv[3])
 #ciklushoz kapcsolódó változók deklarálása
 ciklus=0
@@ -29,7 +29,7 @@ ciklusseged=0
 row=1
 column=0
 while ciklus<idotartam:
-    f.write("%d#" %ciklusseged*nagyt)
+    f.write("%d#" %ciklusseged*int(nagyt))
     f.write("%d#" %(sensor.read_altitude()))
     f.write("%d##\r\n"%(sensor.read_pressure()))
     worksheet.write(row,0,ciklusseged*nagyt)
